Fix build badge in the README
This patch updates the link used in the build badge readme to follow the
new format required by shields.io. Note that the badge now only reflects
the status of the unit_tests.yml workflow which presumably was not the
case when referencing by title was still possible (the title probably
matched ui and unit tests).

See https://github.com/badges/shields/issues/8671