chore: Remove the matrix_sdk_crypto_sync log target
This target used a filter definition that would target only a certain
span. It seems that the `matrix_sdk_crypto` part of the filter
definition wasn't respected so the TRACE log level was enabled for sled
as well.

This would lead to enormous spam in the logs, thus disable it again.