Fix #3881, fix #3893 - Delay the presentation of the share extension room selection screen until the currently presented room is correctly torn down
If QuickLook is presented (e.g. a timeline media is being viewed in full screen) then its animated dismissal will
interfere with presenting other screens and break the state machine. Unfortunately I wasn't able to find any
reasonable way to have the QuickLook dismiss itself without an animation

On the other hand, we can fix both this issue and improve #3893 by just delaying the presentation of the room
selection list, which is what this patch does.