Attempt to fix the wrong timeline start display by defaulting the backwards publisher to .idle
The default values shouldn't matter as they should be updated when calling subscribeToPagination
but empirically we randomly see the timeline start virtual item when we shouldn't.
We believe there's a race condition between the default values the status publisher
so we're going to default the backwards one to .idle.
Worst case it's going to do one extra back pagination.