Update Localizable.strings
"Sie" is the formal way of referring to someone while "Du" is the personal/casual way of talking to someone.

For example, you would say "Du" when youre talking with a friend. 
You would use "Sie" when you try to get a credit from the bank.

Since the new update where the users message is prefixed with the "common_you" in the chat preview, my messages are prefixed with the very formal way which feels cold and not really personal since these messages are my own.... It does feel ok to use the formal way when explaining settings but it feels off when referring to my own actions/content.

For reference, both signal and whatsapp use the casual "Du" in german.