fix intermittent glitches when sending messages.
When UITableView inserts a new row at precisely the same time as us hiding the
Send Receipts on a bubble (due to moving them to the new row), there's a race
which can cause UITableView to miscalculate y-offsets due to the height of the
bubble suddenly changing. To break the race, we hide the SR shortly after the
row is inserted.

Fixes the problem shown in
https://github.com/element-hq/element-x-ios/issues/4127#issuecomment-2889131459